fix: limit memory usage
instead of loading cached objects into memory while processing new data, we load all the pending changes and then write them all out at the end. this means that the moment the cached data comes into memory it is written back out, which limits the amount of time the data is in memory and gives gc more opportunities to collect it.